Update : fixed bitmaps of blasts' circle , lower cost of show Update : linked new game State over Bug fixed : bitmaps of buttons and star show overlaply Bug fixed : lag when candies drop on starup Bug fixed : ending bonus showed too early Bug fixed : candies did't swap when both are powered candy Bug fixed : one more col / row be remove when swap two stripped candy with some condition 3h

+++ game4.10.1/Source/Stage.cpp
@@ -103,7 +103,7 @@ bool game_framework::Stage::IsUnlock()
 {
 	return isUnlock == 1 ? true : false;
 }
-double game_framework::Stage::GetCurrentScore()
+int game_framework::Stage::GetCurrentScore()
 {
 	return currentScore;
 }
